# src/app/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "QR Generator API",
    "version": "1.0.0",
    "description": "REST API for managing QR codes, scans, and webhooks",
    "contact": {
      "name": "API Support",
      "url": "https://your-domain.com/support"
    }
  },
  "servers": [
    {
      "url": "https://your-domain.com/api/v1",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "QRCode": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "url": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "foregroundColor": { "type": "string" },
          "backgroundColor": { "type": "string" },
          "dotType": { "type": "string", "enum": ["square", "rounded", "dots", "classy", "classy-rounded", "extra-rounded"] },
          "cornerType": { "type": "string", "enum": ["square", "rounded", "dot", "extra-rounded"] },
          "logoUrl": { "type": "string" },
          "hasWatermark": { "type": "boolean" },
          "isDynamic": { "type": "boolean" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "maxScans": { "type": "integer" },
          "redirectUrl": { "type": "string" },
          "webhookUrl": { "type": "string" },
          "scanCount": { "type": "integer" },
          "downloadCount": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Scan": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "qrCodeId": { "type": "string" },
          "scannedAt": { "type": "string", "format": "date-time" },
          "userAgent": { "type": "string" },
          "ipAddress": { "type": "string" },
          "country": { "type": "string" },
          "city": { "type": "string" },
          "device": { "type": "string" },
          "browser": { "type": "string" },
          "os": { "type": "string" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" }
        }
      }
    }
  },
  "paths": {
...
    "/qr-codes/{id}": {
      "get": {
        "summary": "Get QR code by ID",
        "tags": ["QR Codes"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "QR code details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "qrCode": { "$ref": "#/components/schemas/QRCode" }
                  }
                }
              }
            }
          },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "put": {
        "summary": "Update QR code",
        "tags": ["QR Codes"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": { "type": "string" },
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "foregroundColor": { "type": "string" },
                  "backgroundColor": { "type": "string" },
                  "dotType": { "type": "string" },
                  "cornerType": { "type": "string" },
                  "logoUrl": { "type": "string" },
                  "hasWatermark": { "type": "boolean" },
                  "isDynamic": { "type": "boolean" },
                  "expiresAt": { "type": "string", "format": "date-time" },
                  "maxScans": { "type": "integer" },
                  "redirectUrl": { "type": "string" },
                  "webhookUrl": { "type": "string" },
                  "isActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "QR code updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "qrCode": { "$ref": "#/components/schemas/QRCode" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete QR code",
        "tags": ["QR Codes"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "QR code deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/qr-codes/{id}/scans": {
      "get": {
        "summary": "List scans for a QR code",
        "tags": ["Scans"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "startDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "endDate", "in": "query", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "List of scans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scans": { "type": "array", "items": { "$ref": "#/components/schemas/Scan" } },
                    "total": { "type": "integer" },
                    "limit": { "type": "integer" },
                    "offset": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/usage": {
      "get": {
        "summary": "Get API usage statistics",
        "tags": ["Usage"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "apiKeyId", "in": "query", "schema": { "type": "string" } },
          { "name": "startDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "endDate", "in": "query", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "Usage statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usage": {
                      "type": "object",
                      "properties": {
                        "requestCount": { "type": "integer" },
                        "errorCount": { "type": "integer" },
                        "avgResponseTime": { "type": "number" },
                        "totalRequestSize": { "type": "integer" },
                        "totalResponseSize": { "type": "integer" }
                      }
                    },
                    "period": {
                      "type": "object",
                      "properties": {
                        "startDate": { "type": "string", "format": "date-time" },
                        "endDate": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}